Predictions/requirements.txt
easyapi==0.1.5
fastapi==0.78.0
fastapi_mail==1.0.9
python-jose==3.3.0
pandas==1.3.4
passlib==1.7.4
pydantic==1.9.1
pymongo==4.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
PyYAML==6.0
starlette==0.19.1
uvicorn==0.17.6
requests==2.28.0
redis==4.3.3
bcrypt
cryptography==3.4.8
PyJWT